feat: Add a few sensors for i35 humidifier. (#14)
* Add a few sensors for i35 humidifier.

There are more, but I feel maybe we can update how these properties
are created (e.g. using a Python descriptor or a helper method, or a
metaclass) before typing more of them.

Also it seems like we could generate a device object entirely from the
configuration json property. Some of the current attributes clearly do
not apply to the humidifier -- and the humidifier attributes do not
apply to the other device types.

* Report wick drying mode.

* Allow wickdry state to False.

This is how the app turns off the timer for drying the wick.

* fetch device type name.